#714  trace api refactoring  - remove unused class    StackFrame, RootStackFrame, SpanEventStackFrame    replaced SpanRecoder, SpanEventRecoder  - added FrameAttachment
