+++ bootstrap/src/main/java/com/navercorp/pinpoint/bootstrap/context/SpanRecorder.java
@@ -5,7 +5,7 @@ import com.navercorp.pinpoint.common.trace.AnnotationKey;
 import com.navercorp.pinpoint.common.trace.LoggingInfo;
 import com.navercorp.pinpoint.common.trace.ServiceType;
 
-public interface SpanRecorder {
+public interface SpanRecorder extends FrameAttachment {
 
     boolean canSampled();
