+++ profiler/src/main/java/com/navercorp/pinpoint/profiler/context/Span.java
@@ -16,6 +16,7 @@
 
 package com.navercorp.pinpoint.profiler.context;
 
+import com.navercorp.pinpoint.bootstrap.context.FrameAttachment;
 import com.navercorp.pinpoint.bootstrap.context.SpanId;
 import com.navercorp.pinpoint.bootstrap.context.TraceId;
 import com.navercorp.pinpoint.common.util.TransactionIdUtils;
@@ -28,8 +29,9 @@ import com.navercorp.pinpoint.thrift.dto.TSpan;
  * @author netspider
  * @author emeroad
  */
-public class Span extends TSpan {
+public class Span extends TSpan implements FrameAttachment {
     private boolean timeRecording = true;
+    private Object frameObject;
     
     public Span() {
     }
@@ -106,5 +108,24 @@ public class Span extends TSpan {
 
     public void setTimeRecording(boolean timeRecording) {
         this.timeRecording = timeRecording;
+    }
+
+    @Override
+    public Object attachFrameObject(Object attachObject) {
+        final Object before = this.frameObject;
+        this.frameObject = attachObject;
+        return before;
+    }
+
+    @Override
+    public Object getFrameObject() {
+        return this.frameObject;
+    }
+
+    @Override
+    public Object detachFrameObject() {
+        final Object delete = this.frameObject;
+        this.frameObject = null;
+        return delete;
     }
 }
