+++ profiler/src/main/java/com/navercorp/pinpoint/profiler/context/SpanEvent.java
@@ -16,6 +16,7 @@
 
 package com.navercorp.pinpoint.profiler.context;
 
+import com.navercorp.pinpoint.bootstrap.context.FrameAttachment;
 import com.navercorp.pinpoint.thrift.dto.TIntStringValue;
 import com.navercorp.pinpoint.thrift.dto.TSpanEvent;
 
@@ -25,11 +26,12 @@ import com.navercorp.pinpoint.thrift.dto.TSpanEvent;
  * @author netspider
  * @author emeroad
  */
-public class SpanEvent extends TSpanEvent {
+public class SpanEvent extends TSpanEvent implements FrameAttachment {
 
     private final Span span;
     private int stackId;
     private boolean timeRecording = true;
+    private Object frameObject;
 
     public SpanEvent(Span span) {
         if (span == null) {
@@ -93,5 +95,24 @@ public class SpanEvent extends TSpanEvent {
 
     public void setTimeRecording(boolean timeRecording) {
         this.timeRecording = timeRecording;
+    }
+
+    @Override
+    public Object attachFrameObject(Object attachObject) {
+        final Object before = this.frameObject;
+        this.frameObject = attachObject;
+        return before;
+    }
+
+    @Override
+    public Object getFrameObject() {
+        return this.frameObject;
+    }
+
+    @Override
+    public Object detachFrameObject() {
+        final Object delete = this.frameObject;
+        this.frameObject = null;
+        return delete;
     }
 }
